Move process arguments handling to renderSuite.run
